requirements.txt
sentence-transformers>=2.7.0
chromadb>=0.5.0
langchain>=0.2.16
langchain-community>=0.2.16
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
langchain-core>=0.2.38
groq>=0.9.0
streamlit>=1.36.0
yfinance>=0.2.43
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.22.0
python-dotenv>=1.0.1
Pillow>=10.4.0
